Added 'Project management' section in 'Project settings'
