+++ collect_app/src/main/res/xml/admin_preferences.xml
@@ -9,32 +9,6 @@
         app:iconSpaceReserved="false">
     </Preference>
 
-    <PreferenceCategory android:title="@string/project_configuration"
-        app:allowDividerAbove="false"
-        app:allowDividerBelow="false"
-        app:iconSpaceReserved="false">
-
-        <PreferenceScreen
-            android:icon="@drawable/ic_outline_qr_code_scanner_24"
-            android:key="import_settings"
-            android:title="@string/reconfigure_with_qr_code_settings_title"
-            android:summary="@string/reconfigure_with_qr_code_settings_subtext"/>
-
-        <org.odk.collect.android.preferences.dialogs.ResetDialogPreference
-            android:icon="@drawable/ic_outline_refresh_24"
-            android:dialogTitle="@string/reset_settings_dialog_title"
-            android:key="reset_settings"
-            android:negativeButtonText="@string/cancel"
-            android:positiveButtonText="@string/reset_settings_button_reset"
-            android:title="@string/reset_project_settings_title"
-            android:summary="@string/reset_project_settings_subtext" />
-
-        <PreferenceScreen
-            android:icon="@drawable/ic_outline_delete_24"
-            android:key="delete_project"
-            android:title="@string/delete_project" />
-    </PreferenceCategory>
-
     <PreferenceCategory android:title="@string/user_access_control"
         app:allowDividerAbove="false"
         app:allowDividerBelow="false"
