+++ collect_app/src/main/res/xml/project_management_preferences.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+    android:title="@string/project_management_section_title">
+
+    <Preference
+        android:icon="@drawable/ic_outline_qr_code_scanner_24"
+        android:key="import_settings"
+        android:title="@string/reconfigure_with_qr_code_settings_title"
+        android:summary="@string/reconfigure_with_qr_code_settings_subtext"/>
+
+    <org.odk.collect.android.preferences.dialogs.ResetDialogPreference
+        android:icon="@drawable/ic_outline_refresh_24"
+        android:dialogTitle="@string/reset_settings_dialog_title"
+        android:key="reset_settings"
+        android:negativeButtonText="@string/cancel"
+        android:positiveButtonText="@string/reset_settings_button_reset"
+        android:title="@string/reset_project_settings_title"
+        android:summary="@string/reset_project_settings_subtext" />
+
+    <Preference
+        android:icon="@drawable/ic_outline_delete_24"
+        android:key="delete_project"
+        android:title="@string/delete_project" />
+</PreferenceScreen>
