+++ strings/src/main/res/values-cs/strings.xml
@@ -694,7 +694,6 @@
   <string name="found_in_main">Zobrazit v Hlavní nabídce</string>
   <string name="found_in_settings">Zobrazit v Obecném nastavení</string>
   <string name="form_entry">Odškrtnout, chcete-li skrýt z nabídky Formulářů</string>
-  <string name="project_configuration">Konfigurace projektu</string>
   <string name="reset_project_settings_title">Resetovat aplikaci…</string>
   <string name="select_all">Vybrat vše</string>
   <string name="clear_all">Vymazat vše</string>
