+++ strings/src/main/res/values-fr/strings.xml
@@ -688,7 +688,6 @@
   <string name="found_in_main">Désélectionner pour cacher sur le menu principal</string>
   <string name="found_in_settings">Désélectionner pour cacher dans les Paramètres</string>
   <string name="form_entry">Décochez pour cacher depuis la saisie du formulaire</string>
-  <string name="project_configuration">Configuration du projet</string>
   <string name="reset_project_settings_title">Réinitialiser l\'application …</string>
   <string name="select_all">Tout sélectionner</string>
   <string name="clear_all">Tout désélectionner </string>
